# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Investment Analyzer API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AnalysisInputDocumentType =
  (typeof AnalysisInputDocumentType)[keyof typeof AnalysisInputDocumentType];

export const AnalysisInputDocumentType = {
  earnings_transcript: "earnings_transcript",
  earnings_presentation: "earnings_presentation",
  s1_filing: "s1_filing",
} as const;

export interface AnalysisInput {
  /** The raw text of the earnings transcript, earnings presentation, or S1 filing */
  documentText: string;
  documentType: AnalysisInputDocumentType;
  /**
   * Optional company name override
   * @nullable
   */
  companyName?: string | null;
}

/**
 * Whether this is good or bad for the company
 */
export type ChecklistItemVerdict =
  (typeof ChecklistItemVerdict)[keyof typeof ChecklistItemVerdict];

export const ChecklistItemVerdict = {
  positive: "positive",
  negative: "negative",
  neutral: "neutral",
  mixed: "mixed",
} as const;

export interface ChecklistItem {
  /** Name of the metric or item */
  metric: string;
  /** Category such as Revenue, Profitability, Growth, Risk, etc. */
  category: string;
  /** Plain English explanation of what this metric is */
  whatItIs: string;
  /** What this metric tells you about the company */
  whatItMeans: string;
  /** Whether this is good or bad for the company */
  verdict: ChecklistItemVerdict;
  /** Brief explanation of why the verdict is what it is */
  verdictReason: string;
  /**
   * Current reported value, if available
   * @nullable
   */
  value?: string | null;
  /**
   * Year-over-year change, if available
   * @nullable
   */
  yoyChange?: string | null;
  /**
   * Quarter-over-quarter change, if available
   * @nullable
   */
  qoqChange?: string | null;
  /**
   * Any red flags or green flags to note
   * @nullable
   */
  flags?: string | null;
}

export type AnalysisDocumentType =
  (typeof AnalysisDocumentType)[keyof typeof AnalysisDocumentType];

export const AnalysisDocumentType = {
  earnings_transcript: "earnings_transcript",
  earnings_presentation: "earnings_presentation",
  s1_filing: "s1_filing",
} as const;

export type AnalysisOverallVerdict =
  (typeof AnalysisOverallVerdict)[keyof typeof AnalysisOverallVerdict];

export const AnalysisOverallVerdict = {
  positive: "positive",
  negative: "negative",
  neutral: "neutral",
  mixed: "mixed",
} as const;

export interface Analysis {
  id: number;
  companyName: string;
  documentType: AnalysisDocumentType;
  checklist: ChecklistItem[];
  /** Overall plain-English summary for a beginner investor */
  summary: string;
  overallVerdict: AnalysisOverallVerdict;
  keyStrengths?: string[];
  keyRisks?: string[];
  createdAt: string;
}

export type AnalysisSummaryDocumentType =
  (typeof AnalysisSummaryDocumentType)[keyof typeof AnalysisSummaryDocumentType];

export const AnalysisSummaryDocumentType = {
  earnings_transcript: "earnings_transcript",
  earnings_presentation: "earnings_presentation",
  s1_filing: "s1_filing",
} as const;

export type AnalysisSummaryOverallVerdict =
  (typeof AnalysisSummaryOverallVerdict)[keyof typeof AnalysisSummaryOverallVerdict];

export const AnalysisSummaryOverallVerdict = {
  positive: "positive",
  negative: "negative",
  neutral: "neutral",
  mixed: "mixed",
} as const;

export interface AnalysisSummary {
  id: number;
  companyName: string;
  documentType: AnalysisSummaryDocumentType;
  overallVerdict: AnalysisSummaryOverallVerdict;
  /** @nullable */
  summary?: string | null;
  createdAt: string;
}

export type AnalysisStatsByDocumentType = {
  earnings_transcript: number;
  earnings_presentation: number;
  s1_filing: number;
};

export type AnalysisStatsByVerdict = {
  positive: number;
  negative: number;
  neutral: number;
  mixed: number;
};

export interface AnalysisStats {
  totalAnalyses: number;
  byDocumentType: AnalysisStatsByDocumentType;
  byVerdict: AnalysisStatsByVerdict;
}